+++ src/cpx_io/cpx_system/cpx_ap/ap4iol.py
@@ -1,9 +1,9 @@
 """CPX-AP-4IOL module implementation"""
 
-import math
-
 from cpx_io.utils.logging import Logging
 from cpx_io.cpx_system.cpx_base import CpxBase
+
+from cpx_io.utils.helpers import div_ceil
 
 from cpx_io.cpx_system.cpx_ap.cpx_ap_module import CpxApModule  # pylint: disable=E0611
 
@@ -23,8 +23,8 @@ class CpxAp4Iol(CpxApModule):
         self.output_register = self.base.next_output_register
         self.input_register = self.base.next_input_register
 
-        self.base.next_output_register += math.ceil(self.information["Output Size"] / 2)
-        self.base.next_input_register += math.ceil(self.information["Input Size"] / 2)
+        self.base.next_output_register += div_ceil(self.information["Output Size"], 2)
+        self.base.next_input_register += div_ceil(self.information["Input Size"], 2)
 
         Logging.logger.debug(
             f"Configured {self} with output register {self.output_register} and input register {self.input_register}"
@@ -65,7 +65,7 @@ class CpxAp4Iol(CpxApModule):
         """read all IO-Link input data
         register order is [msb, ... , ... , lsb]
         """
-        module_input_size = math.ceil(self.information["Input Size"] / 2) - 2
+        module_input_size = div_ceil(self.information["Input Size"], 2) - 2
 
         data = self.base.read_reg_data(self.input_register, length=module_input_size)
         data = [
@@ -96,7 +96,7 @@ class CpxAp4Iol(CpxApModule):
         """set one channel to list of uint16 values
         channel order is [0, 1, 2, 3]
         """
-        module_output_size = math.ceil(self.information["Output Size"] / 2)
+        module_output_size = div_ceil(self.information["Output Size"], 2)
         channel_size = (module_output_size) // 4
 
         register_data = [