+++ src/cpx_io/utils/helpers.py
@@ -1,0 +1,6 @@
+"""Helper functions"""
+
+
+def div_ceil(x: int, y: int) -> int:
+    """Divides two integers and returns the ceiled result"""
+    return (x + y - 1) // y