+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 
 ### Added
+- [Minor] Added utils/helpers with ceil function
 - [Minor] Unittests for cpxe and cpxap modules
 - [Minor] Added pylintrc file to disable f-string highlighting
 - [Minor] CPXAP: Added read function for AP parameters